% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelMatrix.R
\name{getModMatrix}
\alias{getModMatrix}
\title{Prepares data for the main MDC workhorse function}
\usage{
getModMatrix(f, ncells)
}
\arguments{
\item{f}{an R formula object, without a response.  I.e., of the form
"~ x1 + x2 + ..."}

\item{ncells}{number of cells in study area = row dimension of matrices}
}
\value{
A matrix with variables in formula as columns.  Also included is
the names of variables in the formula and whether the model has an intercept.
}
\description{
Function to return the model matrix from a formula object.  I.e., convert
}
\author{
John Lombardi (email: \code{jlombardi@west-inc.com})
}
