% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihoodMDC}
\alias{likelihoodMDC}
\title{Likelihood calculation forthe RSF fitting procedure}
\usage{
likelihoodMDC(beta, X1, X2, locations, k1, k2, maxLagArg)
}
\arguments{
\item{beta}{current values of the coefficients, size k1+k2}

\item{X1}{2-D matric of covariates for RSF model. Size should be ncells x k1}

\item{X2}{2-D matrix of covariates for the probability ofdetection. Size is
ncells x k2}

\item{locations}{Observation matrix. The size is the number of scheduled fixes by 2}

\item{k1}{Number of coefficients in the RSF model}

\item{k2}{Number of coefficients in the probability of detection model}

\item{maxLagArg}{Maximum number of matrix multiplications to try.}
}
\value{
This function returns a likelihood calculation
}
\description{
This is a function to calculate the likelihood. The function is used
in an iterative optimization process. At the end of the process, the likelihood, when
maximized, will provide thevalues for the coefficients from the RSF fitting.
}
\author{
John Lombardi (email: \code{jlombardi@west-inc.com})
}
