% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodRSF.R
\name{likelihoodMDCRSF}
\alias{likelihoodMDCRSF}
\title{Likelihood calculation forthe RSF fitting procedure}
\usage{
likelihoodMDCRSF(beta, X1, locations, k1)
}
\arguments{
\item{beta}{current values of the coefficients, size k1+k2}

\item{X1}{2-D matric of covariates for RSF model. Size should be ncells x k1}

\item{locations}{Observation matrix. The size is the number of scheduled fixes by 2}

\item{k1}{Number of coefficients in the RSF model}
}
\value{
This function returns a likelihood calculation
}
\description{
This is a function to calculate the likelihood. The function is used
in an iterative optimization process. At the end of the process, the likelihood, when
maximized, will provide thevalues for the coefficients from the RSF fitting. This is used when
we do NOT have a probability of detection argument (100% fix success)
}
\author{
John Lombardi (email: \code{jlombardi@west-inc.com})
}
