% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDC.R
\name{MDC}
\alias{MDC}
\title{Fits the discrete choise RSF}
\usage{
MDC(selection, p, locations, ncells, maxLagArg, RSFCovar, LogCovar)
}
\arguments{
\item{selection}{model formula for rsf parameters (the relative probability of
selection parameters).
Covariates in this model are matrices of size ncells X ncells.
Value of element [i,j]
of covariate k in this formula is the kth covariate value for the jth cell.
In this formulation the rsf parameters are not allowed to vary over time.  Any
distance related covariate (e.g, distance from previous cell) should be a matrix with
value of element [i,j] being the distance between the ith and jth cells.
No intercept for this model.}

\item{p}{model formula for probability of detection given presence parameters.  Covariates
in this mocel are vectors of size ncells.  Value in element [i] of variable
k is value of kth covariate for ith site.  Intercept (1) must be present.}

\item{locations}{Cell ID # where the animal was located
(=NA if the animal was not found) during each fix attempt. Cell ID
should be numeric, the minimum possible is 1, and the max = ncells.}

\item{ncells}{number of available cells in the study area.}

\item{maxLagArg}{Integer amount of maximum number of matrix multiplications to perform.}

\item{RSFCovar}{Characer string for RSF covariates}

\item{LogCovar}{Character vector for the covariates for logistic regression}
}
\value{
This function returns a MDC fit.
}
\description{
This is a function to fit the modified
discrete choice RSF to animal locations
when there is <100% fix success, assuming fix success is a function of habitat.
}
\author{
John Lombardi (email: \code{jlombardi@west-inc.com})
}
