% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCWrapper.R
\name{MDCWrapper}
\alias{MDCWrapper}
\title{Prepares data for the main MDC workhorse function}
\usage{
MDCWrapper(habitatDF, CellID, nCells, rsfCovars = NULL,
  probDetCovars = NULL, distColumns = NULL, maxLagArg = 3)
}
\arguments{
\item{habitatDF}{Dataframe of habitat for the area of interest. Columns should include: coordinates,
a unique identifier, and relevant environmental or other covariates for each unit.}

\item{CellID}{Dataframe of fixes for one animal. NOTE: For >1 animals this function will have to be
subsetted to one individual animal. Columns should include:a unique identifier that matches to the
habitatDF, a record of whether or not a fix was successful.}

\item{nCells}{An integer value for the number of cells in the study area.}

\item{rsfCovars}{A vector of strings that the RSF will be regressed on.
The values of the vector needs to match the names of columns in
the \code{habitatDF}. Example inputs: \code{c("elevation","percent.sage")}, \code{c("elevation")}.
As said before, these need to match the \code{colnames(habitatDF)}. If no environmental covariates are wanted,
e.g. an intercept-only model is the RSF model, don't provide any input to rsfCovars argument.}

\item{probDetCovars}{A vector of strings for the probability of detection. The values of the vector needs to
match the names of columns in the \code{habitatDF}. Example inputs: \code{c("elevation","percent.sage")},
\code{c("elevation")}. As said before,
these need to match the \code{colnames(habitatDF)}. If the probability of detection will be an intercept-only
model, don't provide any input to the probDetCovars argument.}

\item{distColumns}{A vector of strings for the name of the columns that contain coordinates in the
\code{habitatDF}. If distance is to be included as a covariate in the RSF, an extra step is needed
to calculate a distance matrix.}

\item{maxLagArg}{An integer for the number of max matrix multiplications to perform. The default is 3.
Anything larger than 3 will greatly increase computation time.}
}
\value{
This function returns a MDC fit.
}
\description{
This function takes inputs for locations, habitat, and covariates for the RSF and probability
of detection. A bit of data massaging goes on to convert dataframes to a more usable format for the
main MDC function.
}
\author{
John Lombardi (email: \code{jlombardi@west-inc.com})
}
